fix: id type didn't match everywhere So the entity, repo and controller now uses long

+++ src/main/java/se/experis/academy/diary_app/repository/EntryRepository.java
@@ -8,11 +8,11 @@ import se.experis.academy.diary_app.model.Entry;
 import java.util.List;
 
 @Repository
-public interface EntryRepository extends CrudRepository<Entry, Integer> {
+public interface EntryRepository extends CrudRepository<Entry, Long> {
 
     List<Entry> findByActiveTrue();
 
-    Entry findById(int id);
+    Entry findById(long id);
 
 
     @Override
